// packages/adapter-http/generated/schema.ts
export interface paths {
  '/api/v1/meta/version': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 获取数据版本信息 */
    get: operations['MetaController_getVersion_v1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/api/v1/oier/{uid}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * 获取选手详细信息
     * @description 根据选手 ID 获取选手的详细信息
     */
    get: operations['OIerController_getOIer_v1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/api/v1/oier': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * 获取选手列表
     * @description 根据条件查询选手列表，支持分页
     */
    get: operations['OIerController_listOIers_v1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/api/v1/contest/{id}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * 获取比赛详细信息
     * @description 根据比赛 ID 获取比赛的详细信息及参赛选手
     */
    get: operations['ContestController_getContest_v1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/api/v1/contest': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * 获取比赛列表
     * @description 根据条件查询比赛列表，支持分页
     */
    get: operations['ContestController_listContests_v1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/api/v1/school/{id}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * 获取学校详细信息
     * @description 根据学校 ID 获取学校的详细信息
     */
    get: operations['SchoolController_getSchool_v1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/api/v1/school': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /**
     * 获取学校列表
     * @description 根据条件查询学校列表，支持分页
     */
    get: operations['SchoolController_listSchools_v1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
}
// ...
export interface operations {
  MetaController_getVersion_v1: {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 数据版本信息 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['VersionResponseDto'];
        };
      };
    };
  };
  OIerController_getOIer_v1: {
    parameters: {
      query?: never;
      header?: never;
      path: {
        /** @description 选手 ID */
        uid: number;
      };
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 成功返回选手信息 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['GetOIerResponseDto'];
        };
      };
      /** @description 选手不存在 */
      404: {
        headers: {
          [name: string]: unknown;
        };
        content?: never;
      };
    };
  };
  OIerController_listOIers_v1: {
    parameters: {
      query?: {
        /** @description 姓名 */
        name?: string;
        /** @description 姓名拼音首字母（小写） */
        initials?: string;
        /** @description 初中入学年份 */
        enroll_middle?: number;
        /** @description 性别（-1: 女；0: 未知；1: 男） */
        gender?: components['schemas']['Gender'];
        /** @description 省份 */
        province?: string;
        /** @description 页码 */
        page?: number;
        /** @description 每页数量 */
        perPage?: number;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 成功返回选手列表 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['ListOIersResponseDto'];
        };
      };
    };
  };
  ContestController_getContest_v1: {
    parameters: {
      query?: {
        /** @description 页码 */
        page?: number;
        /** @description 每页数量 */
        perPage?: number;
      };
      header?: never;
      path: {
        /** @description 比赛 ID */
        id: number;
      };
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 成功返回比赛信息 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['GetContestResponseDto'];
        };
      };
      /** @description 比赛不存在 */
      404: {
        headers: {
          [name: string]: unknown;
        };
        content?: never;
      };
    };
  };
  ContestController_listContests_v1: {
    parameters: {
      query?: {
        /** @description 比赛类型 */
        type?: string;
        /** @description 比赛年份 */
        year?: number;
        /** @description 页码 */
        page?: number;
        /** @description 每页数量 */
        perPage?: number;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 成功返回比赛列表 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['ListContestsResponseDto'];
        };
      };
    };
  };
  SchoolController_getSchool_v1: {
    parameters: {
      query?: never;
      header?: never;
      path: {
        /** @description 学校 ID */
        id: number;
      };
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 成功返回学校信息 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['GetSchoolResponseDto'];
        };
      };
      /** @description 学校不存在 */
      404: {
        headers: {
          [name: string]: unknown;
        };
        content?: never;
      };
    };
  };
  SchoolController_listSchools_v1: {
    parameters: {
      query?: {
        /** @description 学校名称 */
        name?: string;
        /** @description 省份 */
        province?: string;
        /** @description 城市 */
        city?: string;
        /** @description 页码 */
        page?: number;
        /** @description 每页数量 */
        perPage?: number;
      };
      header?: never;
      path?: never;
      cookie?: never;
    };
    requestBody?: never;
    responses: {
      /** @description 成功返回学校列表 */
      200: {
        headers: {
          [name: string]: unknown;
        };
        content: {
          'application/json': components['schemas']['ListSchoolsResponseDto'];
        };
      };
    };
  };
}